backend/pyproject.toml
[tool.poetry]
name = "codesentinel"
version = "0.1.0"
description = "LLM-powered automated PR code reviewer"
authors = ["CodeSentinel Team"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic-settings = "^2.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.0"
qdrant-client = "^1.9.0"
sentence-transformers = "^3.0.0"
langchain = "^0.2.0"
langchain-community = "^0.2.0"
PyGithub = "^2.3.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
structlog = "^24.2.0"
PyYAML = "^6.0.0"
torch = { version = "2.2.2", source = "pytorch" }
setuptools = ">=70.0.0"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-instrumentation-celery = "^0.46b0"
opentelemetry-instrumentation-sqlalchemy = "^0.46b0"
opentelemetry-instrumentation-redis = "^0.46b0"
opentelemetry-instrumentation-httpx = "^0.46b0"
opentelemetry-exporter-otlp = "^1.25.0"
langfuse = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
anyio = "^4.4.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
